--- a/tests/SearchableScopeTest.php
+++ b/tests/SearchableScopeTest.php
@@ -1,0 +1,33 @@
+<?php
+
+namespace Laravel\Scout\Tests;
+
+use Illuminate\Database\Eloquent\Builder;
+use Laravel\Scout\SearchableScope;
+use Mockery as m;
+use PHPUnit\Framework\TestCase;
+
+class SearchableScopeTest extends TestCase
+{
+    public function tearDown()
+    {
+        m::close();
+    }
+
+    public function test_chunks_by_id()
+    {
+        $builder = m::mock(Builder::class);
+        $builder->shouldReceive('macro')->with('searchable', m::on(function ($callback) use ($builder) {
+            $builder->shouldReceive('chunkById')->with(500, m::type(\Closure::class));
+            $callback($builder, 500);
+            return true;
+        }));
+        $builder->shouldReceive('macro')->with('unsearchable', m::on(function ($callback) use ($builder) {
+            $builder->shouldReceive('chunkById')->with(500, m::type(\Closure::class));
+            $callback($builder, 500);
+            return true;
+        }));
+
+        (new SearchableScope())->extend($builder);
+    }
+}