Using chunkById is dramatically faster when importing millions of rows.  This issue was originally reported in #182.

On a table with 1,632,576 rows, using the null scout driver so we don't have any noise from HTTP requests, an import goes from 29 minutes and 57 seconds to 28 seconds.

Before:

```
$ time php artisan scout:import "App\User"
All [App\User] records have been imported.
php artisan scout:import "App\User"  14.80s user 3.69s system 1% cpu 29:57.14 total
```

After:

```
$ time php artisan scout:import "App\User"
php artisan scout:import "App\User"  17.18s user 4.06s system 73% cpu 28.809 total
```

Concerns:

`chunkById` currently removes all ORDER BY clauses.  If someone had called `searchable` like this it wouldn't work as expected anymore:

```php
App\Order::where('price', '>', 100)->orderBy('price')->searchable();
```

I doubt that anyone actually needs that behavior?  If they do there are a few options:

- Make this opt in with a config setting, i.e. `config('scout.chunk.by_id')`
- Make `chunkById` preserve order by clauses.  This is [explained here](https://use-the-index-luke.com/sql/partial-results/fetch-next-page) but it's a little confusing.  I've implemented this before for a proprietary ElasticSearch indexer.  It shouldn't be too difficult to add to Eloquent.