@driesvints 2019-03-05
Hmm, I agree that the trad-off is definitely worth it. Wondering what @nunomaduro thinks.